+++ django_athm/utils.py
@@ -28,6 +28,35 @@ class BaseHTTPAdapter:
 
 
 class DummyHTTPAdapter(BaseHTTPAdapter):
+    def get_with_data(self, url, data):
+        return [
+            {
+                "transactionType": "refund",
+                "referenceNumber": "212831546-7638e92vjhsbjbsdkjqbjkbqdq",
+                "date": "2019-06-06 17:12:02.0",
+                "refundedAmount": "1.00",
+                "total": "1.00",
+                "tax": "1.00",
+                "subtotal": "1.00",
+                "metadata1": "metadata1 test",
+                "metadata2": "metadata2 test",
+                "items": '[{"name":"First Item","description":"This is a description.","quantity":"1","price":"1.00","tax":"1.00","metadata":"metadata test"},{"name":"Second Item","description":"This is another description.","quantity":"1","price":"1.00","tax":"1.00","metadata":"metadata test"}]',
+            },
+            {
+                "transactionType": "payment",
+                "status": "completed",
+                "referenceNumber": "212831546-402894d56b240610016b2e6c78a6003a",
+                "date": "2019-06-06 16:12:02.0",
+                "refundedAmount": "0.00",
+                "total": "5.00",
+                "tax": "1.00",
+                "subtotal": "4.00",
+                "metadata1": "metadata1 test",
+                "metadata2": "metadata2 test",
+                "items": '[{"name":"First Item","description":"This is a description.","quantity":"1","price":"1.00","tax":"1.00","metadata":"metadata test"},{"name":"Second Item","description":"This is another description.","quantity":"1","price":"1.00","tax":"1.00","metadata":"metadata test"}]',
+            },
+        ]
+
     def get(self, url):
         logger.debug(f"[DummyHTTPAdapter:get] URL: {url}")